Adds question create test
