+++ app/tests/reader_studies_tests/test_permissions.py
@@ -7,25 +7,8 @@ from tests.reader_studies_tests.factories import (
     QuestionFactory,
     AnswerFactory,
 )
-from tests.reader_studies_tests.utils import get_rs_creator
+from tests.reader_studies_tests.utils import get_rs_creator, TwoReaderStudies
 from tests.utils import get_view_for_user
-
-
-class TwoReaderStudies:
-    def __init__(self):
-        self.creator = get_rs_creator()
-        self.rs1, self.rs2 = ReaderStudyFactory(), ReaderStudyFactory()
-        self.editor1, self.reader1, self.editor2, self.reader2 = (
-            UserFactory(),
-            UserFactory(),
-            UserFactory(),
-            UserFactory(),
-        )
-        self.rs1.add_editor(user=self.editor1)
-        self.rs2.add_editor(user=self.editor2)
-        self.rs1.add_reader(user=self.reader1)
-        self.rs2.add_reader(user=self.reader2)
-        self.u = UserFactory()
 
 
 @pytest.mark.django_db
